[#25] Add rubocop config from bundler (#26)

+++ lib/manpages/uninstall.rb
@@ -1,7 +1,5 @@
 module Manpages
-
   class Uninstall
-
     def initialize(gem_spec, gem_dir, target_dir)
       @gem_spec   = gem_spec
       @gem_dir    = gem_dir
@@ -9,10 +7,10 @@ module Manpages
     end
 
     def uninstall_manpages
-      unlink_manpages if GemVersion.new(@gem_spec).is_latest?
+      unlink_manpages if GemVersion.new(@gem_spec).latest?
     end
 
-    private
+  private
 
     def unlink_manpages
       ManFiles.new(@gem_dir, @target_dir).manpages.each do |file|
@@ -22,10 +20,8 @@ module Manpages
 
     def unlink_manpage(file)
       man_target_file = ManFiles.new(@gem_dir, @target_dir).man_file_path(file)
-      if File.symlink?(man_target_file) && File.readlink(man_target_file) == file
-        FileUtils.rm(man_target_file)
-      end
+      FileUtils.rm(man_target_file) if File.symlink?(man_target_file) &&
+          File.readlink(man_target_file) == file
     end
-
   end
 end
